Conflicts:
	opensource/kernel-cpp/scheduler/scheduler.xsd
	opensource/kernel-cpp/scheduler/spooler_embedded_files.cxx
	opensource/kernel-cpp/scheduler/spooler_job.cxx
	opensource/kernel-cpp/scheduler/spooler_log.cxx
	opensource/kernel-cpp/scheduler/spooler_wait.cxx
	opensource/test/src/main/scala/com/sos/scheduler/engine/test/scala/SchedulerTestImplicits.scala
	opensource/test/src/test/java/com/sos/scheduler/engine/tests/jira/js804/JS804IT.java
	opensource/test/src/test/scala/com/sos/scheduler/engine/tests/jira/js803/JS803IT.scala

+++ opensource/kernel-cpp/kram/sos_simple.h
@@ -25,7 +25,7 @@ struct Fill_zero
 
     Fill_zero( const Fill_end& ende )
     {
-        int length = (char*)&ende - (char*)( this+1 );
+        size_t length = (char*)&ende - (char*)( this+1 );
         assert( length >= 0 );
         memset( this + 1, 0, length );
     }
